Move num-threads and cpu-affinity to common opts
The option settings are split into common, server, client (which is
really client common), and client mode specific settings.  Move
the two listed options to the common settings.  This does not make
the options *work* for clients yet, just makes them accepted for
clients.  With the exception of changing the help comment for
num-threads, this made not other change except to move the code
location.

Signed-off-by: Doug Ledford <dledford@redhat.com>